edmund.html update
Added pattern/regex recognition to the input fields to where it won't
allow numbers or special characters in their names and makes it to where
they have to have a troy.edu email address. Also added the green
checkmark and the red X for whether or not the password was confirmed or
not. Only thing we need to change is to make the script a js file and
make a some php for the registration process.